Develop

+++ src/main/java/com/novoa/videogames/service/SaleService.java
@@ -5,6 +5,8 @@ import com.novoa.videogames.entity.Sale;
 import com.novoa.videogames.repository.SaleRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.OptimisticLockingFailureException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.NoSuchElementException;
@@ -17,17 +19,23 @@ public class SaleService {
     public Iterable<Sale> getSaleIterable(){
         return saleRepository.findAll();
     }
-    public Sale getSaleById(String saleId){
+    public ResponseEntity<Sale> getSaleById(String saleId){
+        if(saleExistsById(saleId)){
+            return new ResponseEntity("There is no sale with that Id: " + saleId, HttpStatus.NOT_FOUND);
+        }
         try{
-            return saleRepository.findById(saleId).get();
+            return ResponseEntity.ok(saleRepository.findById(saleId).get());
         } catch (IllegalArgumentException e){
             System.out.println("The given argument was null while trying to fetch it, in Sale Repository");
         } catch (NoSuchElementException e ){
             System.out.println("There was no element with the given Id: "+ saleId +" while trying to fetch by Id Sale");
         }
-        return new Sale();
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
-    public Sale saveSale(SaleDto saleDto){
+    public ResponseEntity<Sale> saveSale(SaleDto saleDto){
+        if(saleExistsById(saleDto.getSaleId())){
+            return new ResponseEntity("The sale Id already exists", HttpStatus.BAD_REQUEST);
+        }
         Sale sale = new Sale(
                 saleDto.getSaleId(),
                 saleDto.getProductName(),
@@ -35,32 +43,40 @@ public class SaleService {
                 saleDto.getDiscount()
         );
         try {
-            return saleRepository.save(sale);
+            return ResponseEntity.ok(saleRepository.save(sale));
         } catch (IllegalArgumentException | OptimisticLockingFailureException e){
             System.out.println("The given argument was null while trying to save, in Sale Repository");
         }
-        return new Sale();
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
-    public Sale updateSale(SaleDto saleDto){
+    public ResponseEntity<Sale> updateSale(SaleDto saleDto){
+        if(!saleExistsById(saleDto.getSaleId())){
+            return new ResponseEntity("The sale with that Id does not exist", HttpStatus.BAD_REQUEST);
+        }
         try {
             Sale sale = saleRepository.findById(saleDto.getSaleId()).get();
             sale.setSaleId(saleDto.getSaleId());
             sale.setProductName(saleDto.getProductName());
             sale.setPrice(saleDto.getPrice());
             sale.setDiscount(saleDto.getDiscount());
-            return saleRepository.save(sale);
+
+            return ResponseEntity.ok(saleRepository.save(sale));
         } catch (IllegalArgumentException | OptimisticLockingFailureException e){
             System.out.println("The given argument was null while trying to save, in Sale Repository");
         } catch (NoSuchElementException e ){
             System.out.println("There was no element with the given Id, while trying to update Sale");
         }
-        return new Sale();
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
-    public void deleteSale(String saleId){
+    public ResponseEntity<String> deleteSale(String saleId){
+        if(!saleExistsById(saleId)){
+            return new ResponseEntity("There is no sale with that Id: " + saleId, HttpStatus.NOT_FOUND);
+        }
         try{
             Sale sale = saleRepository.findById(saleId).get();
             if(!sale.equals(Optional.empty())){
                 saleRepository.deleteById(saleId);
+                return new ResponseEntity("Sale deleted successfully", HttpStatus.OK);
             } else {
                 System.out.println("There is no such element with Id: " + saleId + " In Sale Repository");
             }
@@ -69,6 +85,7 @@ public class SaleService {
         } catch (IllegalArgumentException e ){
             System.out.println("The given argument was null while trying to delete, in Sale Repository");
         }
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
     public boolean saleExistsById(String saleId){
         try{
@@ -86,13 +103,13 @@ public class SaleService {
         }
         return false;
     }
-    public int getDiscountReport(){
+    public ResponseEntity<String> getDiscountReport(){
         int totalDiscount = 0;
         for (Sale sale: this.getSaleIterable()) {
             if (sale.getDiscount()!=null){
                 totalDiscount = totalDiscount + Integer.parseInt(sale.getDiscount());
             }
         }
-        return totalDiscount;
+        return new ResponseEntity<>("The total given by discounts is: " + totalDiscount , HttpStatus.OK);
     }
 }
